fix: DTO 데이터 타입 변경
fix: DTO 데이터 타입 변경

+++ src/main/java/com/team/saver/market/order/dto/OrderMenuResponse.java
@@ -11,14 +11,14 @@ public class OrderMenuResponse {
 
     private String menuName;
 
-    private int price;
+    private long price;
 
     private String optionDescription;
 
-    private int optionPrice;
+    private long optionPrice;
 
-    private int amount;
+    private long amount;
 
-    private int finalPrice;
+    private long finalPrice;
 
 }
